Moved refactored EventCounterParty
